Corrige overflow real: chat_id do Telegram precisa ser BigInteger
Bug encontrado testando o bot de verdade: Victor deu /start no
@veaga_dicas_bot e o webhook devolvia 500 (confirmado via
getWebhookInfo do Telegram). Causa raiz: contas novas do Telegram tem
id maior que 2.147.483.647 (limite de um Integer comum no Postgres),
entao gravar o chat_id real dava overflow. Passou despercebido nos
testes porque SQLite (usado localmente) nao reforca largura de coluna
-- so apareceu contra o Postgres de producao com um chat_id de verdade.

Corrigido: chat_id agora e' BigInteger. Coluna ja migrada em producao
(scripts/migrar_chat_id_bigint.py) e testada com um chat_id real acima
do limite antigo (5123456789) -- grava sem erro agora. Testes novos
cobrem o tipo da coluna e um insert com chat_id grande.

+++ backend/app/models/telegram_subscriber.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, DateTime
+from sqlalchemy import Column, Integer, BigInteger, Boolean, DateTime
 from sqlalchemy.sql import func
 from app.database import Base
 
@@ -10,7 +10,12 @@ class TelegramSubscriber(Base):
     # Identificador do chat no Telegram -- unico por usuario/grupo que deu
     # /start no bot. E' o unico dado que precisamos guardar (o bot nao
     # tem conceito de conta VEAGA nenhuma, so avisa quem deu /start).
-    chat_id = Column(Integer, unique=True, nullable=False, index=True)
+    # BigInteger e' obrigatorio aqui: contas novas do Telegram ja tem id
+    # maior que o limite de um Integer normal (2.147.483.647) -- um
+    # Integer comum causava overflow no Postgres de producao (silencioso
+    # no SQLite local, que nao reforca o tamanho do campo, por isso so
+    # apareceu com um chat_id real).
+    chat_id = Column(BigInteger, unique=True, nullable=False, index=True)
     # False quando o usuario deu /parar (ou bloqueou o bot e o envio
     # devolveu 403) -- nunca deletamos a linha, so paramos de mandar.
     ativo = Column(Boolean, nullable=False, default=True)
